Conflicts:
	krita/plugins/formats/xcf/3rdparty/xcftools/config.h
	krita/ui/CMakeLists.txt
	krita/ui/KisDocumentEntry.h
	krita/ui/KisFilterEntry.h
	krita/ui/canvas/kis_grid_manager.cpp
	krita/ui/input/wintab/kis_tablet_support_x11.h
	libs/db/expression.cpp

+++ krita/ui/kis_paintop_box.h
@@ -37,7 +37,7 @@
 #include "kis_locked_properties_server.h"
 #include "kis_locked_properties.h"
 #include "kis_config.h"
-#include "krita_export.h"
+#include "kritaui_export.h"
 
 
 class QToolButton;
@@ -197,8 +197,8 @@ private:
     QString             m_currCompositeOpID;
     KisNodeWSP          m_previousNode;
 
-    int normalBrushSize; // when toggling between eraser mode
-    int eraserBrushSize;
+    qreal normalBrushSize; // when toggling between eraser mode
+    qreal eraserBrushSize;
 
     KisAction* m_hMirrorAction;
     KisAction* m_vMirrorAction;
